LLM_API/requirements.txt
flask
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
